充值成功，通知钱包换换钻接口。

+++ src/main/java/com/blemobi/payment/rest/common/WalletTools.java
@@ -17,7 +17,7 @@ import lombok.extern.log4j.Log4j;
 @Log4j
 public class WalletTools {
 	//这是支付充传值成功后对钱包系统的通知
-	public static boolean invokeWalletDiamondAdd(String uuid,String token,int rmoney,String ordernumber){
+	public static boolean invokeWalletDiamondAdd(String uuid,String token,long rmoney,String ordernumber){
 		boolean rtn = false;
 		int max=5; //通讯连接，最多重复5遍。
 		while(max-->0){
@@ -35,7 +35,7 @@ public class WalletTools {
 		
 	}
 	
-	private static boolean invokeWalletDiamondAdd2(String uuid,String token,int rmoney,String ordernumber) throws Exception {
+	private static boolean invokeWalletDiamondAdd2(String uuid,String token,long rmoney,String ordernumber) throws Exception {
 		boolean rtn = false;
 		ClientUtilImpl clientUtil = new ClientUtilImpl();
